Separated code into modules
